# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Для некоторых действий требуется авторизация с помощью Bearer Token",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация",
                "description": "Регистрация нового пользователя и возвращает access и refresh токенов. Успешный ответ будет содержать Bearer токен, который нужно использовать для авторизации в других запросах.",
                "operationId": "register",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "status",
                                        "content"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/AuthTokenResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Access token имеет ограниченный срок действия. Для его обновления используйте endpoint /refresh с refresh token.",
                "operationId": "login",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "status",
                                        "content"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/AuthTokenResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "description": "Обновление access token с помощью refresh token.",
                "operationId": "refresh",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful refresh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "status",
                                        "content"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/AuthTokenResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Выход из системы. Удаляет текущий токен доступа пользователя.",
                "operationId": "logout",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessfulEmptyResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout-all": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout для всех устройств",
                "description": "Выход из системы на всех устройствах. Удаляет все токены доступа пользователя.",
                "operationId": "logoutAllSessions",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessfulEmptyResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Подтверждение регистрации",
                "description": "Подтверждение регистрации пользователя с помощью кода верификации.",
                "operationId": "verifyRegistration",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessfulEmptyResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/verify/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Повторная отправка кода верификации",
                "description": "Повторная отправка кода верификации на email пользователя. Код будет отправлен на email, указанный при регистрации.",
                "operationId": "verifyResend",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessfulEmptyResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Отправка ссылки для сброса пароля",
                "description": "Отправка ссылки для сброса пароля на email пользователя.",
                "operationId": "resetPasswordToken",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/XRequestedWithHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessfulEmptyResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "remember": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "AuthTokenResponse": {
                "required": [
                    "accessToken",
                    "refreshToken",
                    "tokenType",
                    "expiresIn"
                ],
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "example": "1|abc..."
                    },
                    "refreshToken": {
                        "type": "string",
                        "example": "2|xyz..."
                    },
                    "tokenType": {
                        "type": "string",
                        "example": "Bearer"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
...
            "SuccessfulEmptyResponse": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "ok"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "XRequestedWithHeader": {
                "name": "X-Requested-With",
                "in": "header",
                "description": "Must be 'XMLHttpRequest'",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "XMLHttpRequest"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}